# Conflicts:
#	SCMM.Azure.ApplicationInsights/Filters/IgnoreStaticWebFilesFilter.cs
#	SCMM.Discord.Bot.Server/Program.cs
#	SCMM.Steam.API/Commands/ImportSteamAppItemDefinitionsArchive.cs
#	SCMM.Steam.Functions/Program.cs
#	SCMM.Web.Client/Pages/Market/MarketActivityPage.razor
#	SCMM.Web.Client/Pages/SystemPage.razor
#	SCMM.Web.Client/Shared/Layout/AppNavigationMenu.razor
#	SCMM.Web.Server/Program.cs

+++ SCMM.Steam.Data.Models/Enums/SteamProfileInventoryItemFlags.cs
@@ -1,7 +1,7 @@
 ﻿namespace SCMM.Steam.Data.Models.Enums
 {
     [Flags]
-    public enum SteamProfileInventoryItemFlags : byte
+    public enum SteamProfileInventoryItemFlags : int
     {
         None = 0x00,
         Investment = 0x01,
